+++ ydb/core/health_check/self_check_request.h
@@ -1,27 +1,22 @@
 #pragma once
 
+#include "builder.h"
 #include "log.h"
 #include "health_check.h"
-#include "self_check_result.h"
+#include "health_check_data.h"
+#include "health_check_helper.h"
 
-#include <ydb/core/base/hive.h>
 #include <ydb/core/base/path.h>
 #include <ydb/core/base/statestorage.h>
-#include <ydb/core/base/tablet_pipe.h>
-#include <ydb/core/cms/console/configs_dispatcher.h>
 #include <ydb/core/mon/mon.h>
 #include <ydb/core/blobstorage/base/blobstorage_events.h>
-#include <ydb/core/cms/console/console.h>
 #include <ydb/core/mind/tenant_slot_broker.h>
-#include <ydb/core/node_whiteboard/node_whiteboard.h>
 #include <ydb/core/tx/scheme_cache/scheme_cache.h>
 #include <ydb/core/tx/schemeshard/schemeshard.h>
 #include <ydb/core/util/proto_duration.h>
 #include <ydb/core/util/tuples.h>
 
-#include <ydb/core/protos/blobstorage_distributed_config.pb.h>
 #include <ydb/core/protos/config.pb.h>
-#include <ydb/core/sys_view/common/events.h>
 #include <ydb/library/actors/interconnect/interconnect.h>
 #include <ydb/library/actors/wilson/wilson_span.h>
 #include <ydb/library/wilson_ids/wilson.h>
@@ -58,8 +53,6 @@ using namespace NSchemeCache;
 using namespace NSchemeShard;
 using namespace NSysView;
 using namespace NConsole;
-using namespace NNodeWhiteboard;
-using NNodeWhiteboard::TTabletId;
 
 void RemoveUnrequestedEntries(Ydb::Monitoring::SelfCheckResult& result, const Ydb::Monitoring::SelfCheckRequest& request) {
     if (!request.return_verbose_status()) {
@@ -126,121 +119,6 @@ public:
     enum ETimeoutTag {
         TimeoutBSC,
         TimeoutFinal,
-    };
-
-    struct TTenantInfo {
-        TString Name;
-        Ydb::Cms::GetDatabaseStatusResult::State State;
-    };
-
-    struct TNodeTabletState {
-        struct TTabletStateSettings {
-            TInstant AliveBarrier;
-            ui32 MaxRestartsPerPeriod; // per hour
-            ui32 MaxTabletIdsStored = 10;
-            bool ReportGoodTabletsIds = false;
-        };
-
-        enum class ETabletState {
-            Good,
-            Stopped,
-            RestartsTooOften,
-            Dead,
-        };
-
-        struct TNodeTabletStateCount {
-            NKikimrTabletBase::TTabletTypes::EType Type;
-            ETabletState State;
-            bool Leader;
-            int Count = 1;
-            TStackVec<TString> Identifiers;
-
-            static ETabletState GetState(const NKikimrHive::TTabletInfo& info, const TTabletStateSettings& settings) {
-                if (info.volatilestate() == NKikimrHive::TABLET_VOLATILE_STATE_STOPPED) {
-                    return ETabletState::Stopped;
-                }
-                ETabletState state = (info.restartsperperiod() >= settings.MaxRestartsPerPeriod) ? ETabletState::RestartsTooOften : ETabletState::Good;
-                if (info.volatilestate() == NKikimrHive::TABLET_VOLATILE_STATE_RUNNING) {
-                    return state;
-                }
-                if (info.tabletbootmode() != NKikimrHive::TABLET_BOOT_MODE_DEFAULT) {
-                    return state;
-                }
-                if (info.lastalivetimestamp() != 0 && TInstant::MilliSeconds(info.lastalivetimestamp()) < settings.AliveBarrier) {
-                    // Tablet is not alive for a long time
-                    // We should report it as dead unless it's just waiting to be created
-                    if (info.generation() == 0 && info.volatilestate() == NKikimrHive::TABLET_VOLATILE_STATE_BOOTING && !info.inwaitqueue()) {
-                        return state;
-                    }
-                    return ETabletState::Dead;
-                }
-                return state;
-
-            }
-
-            TNodeTabletStateCount(const NKikimrHive::TTabletInfo& info, const TTabletStateSettings& settings)
-                : Type(info.tablettype())
-                , State(GetState(info, settings))
-                , Leader(info.followerid() == 0)
-            {
-            }
-
-            bool operator ==(const TNodeTabletStateCount& o) const {
-                return State == o.State && Type == o.Type && Leader == o.Leader;
-            }
-        };
-
-        TStackVec<TNodeTabletStateCount> Count;
-
-        void AddTablet(const NKikimrHive::TTabletInfo& info, const TTabletStateSettings& settings) {
-            TNodeTabletStateCount tabletState(info, settings);
-            auto itCount = Find(Count, tabletState);
-            if (itCount != Count.end()) {
-                itCount->Count++;
-            } else {
-                Count.emplace_back(tabletState);
-                itCount = std::prev(Count.end());
-            }
-            if (itCount->State != ETabletState::Good || settings.ReportGoodTabletsIds) {
-                if (itCount->Identifiers.size() < settings.MaxTabletIdsStored) {
-                    TStringBuilder id;
-                    id << info.tabletid();
-                    if (info.followerid()) {
-                        id << '.' << info.followerid();
-                    }
-                    itCount->Identifiers.emplace_back(id);
-                }
-            }
-        }
-    };
-
-    struct TStoragePoolState {
-        TString Name;
-        THashSet<TGroupId> Groups;
-    };
-
-    struct TDatabaseState {
-        TTabletId HiveId = {};
-        TTabletId SchemeShardId = {};
-        TPathId ResourcePathId = {};
-        TVector<TNodeId> ComputeNodeIds;
-        THashSet<ui64> StoragePools;        // BSConfig case
-        THashSet<TString> StoragePoolNames; // no BSConfig case
-        THashMap<std::pair<TTabletId, TFollowerId>, const NKikimrHive::TTabletInfo*> MergedTabletState;
-        THashMap<TNodeId, TNodeTabletState> MergedNodeTabletState;
-        THashMap<TNodeId, ui32> NodeRestartsPerPeriod;
-        ui64 StorageQuota = 0;
-        ui64 StorageUsage = 0;
-        TMaybeServerlessComputeResourcesMode ServerlessComputeResourcesMode;
-        TNodeId MaxTimeDifferenceNodeId = 0;
-        TString Path;
-    };
-
-    struct TGroupState {
-        TString ErasureSpecies;
-        std::vector<const NKikimrSysView::TVSlotEntry*> VSlots;
-        ui32 Generation;
-        bool LayoutCorrect = true;
     };
 
     template<typename T>
@@ -479,72 +357,9 @@ public:
 
     THashMap<ui64, TRequestResponse<TEvTxProxySchemeCache::TEvNavigateKeySetResult>> NavigateKeySet;
 
-    struct THintOverloadedShard {
-        TTabletId TabletId;
-        TFollowerId FollowerId;
-        double CPUCores;
-        TTabletId SchemeShardId;
-        TString Message;
-    };
-
     THashMap<TString, THintOverloadedShard> OverloadedShardHints;
     static constexpr size_t MAX_OVERLOADED_SHARDS_HINTS = 10;
     static constexpr double OVERLOADED_SHARDS_CPU_CORES = 0.75;
-
-    struct TTabletRequestsState {
-        struct TTabletState {
-            TTabletTypes::EType Type = TTabletTypes::Unknown;
-            TString Database;
-            bool IsUnresponsive = false;
-            TDuration MaxResponseTime;
-            TActorId TabletPipe = {};
-        };
-
-        struct TRequestState {
-            TTabletId TabletId;
-            TString Key;
-            TMonotonic StartTime;
-        };
-
-        std::unordered_map<TTabletId, TTabletState> TabletStates;
-        std::unordered_map<ui64, TRequestState> RequestsInFlight;
-        ui64 RequestId = 0;
-
-        // Some tablets (currently only BSC sys view) do not set response cookie
-        // So we use special constant ids for those requests
-        enum ERequestId : ui64 {
-            RequestStoragePools = 1'000'000,
-            RequestGroups,
-            RequestVSlots,
-            RequestPDisks,
-            RequestGetPartitionStats,
-        };
-
-        void MakeRequest(TTabletId tabletId, const TString& key, ui64 requestId) {
-            RequestsInFlight.emplace(requestId, TRequestState{tabletId, key, TMonotonic::Now()});
-        }
-
-        ui64 MakeRequest(TTabletId tabletId, const TString& key) {
-            MakeRequest(tabletId, key, ++RequestId);
-            return RequestId;
-        }
-
-        TTabletId CompleteRequest(ui64 requestId) {
-            TTabletId tabletId = {};
-            TMonotonic finishTime = TMonotonic::Now();
-            auto itRequest = RequestsInFlight.find(requestId);
-            if (itRequest != RequestsInFlight.end()) {
-                TDuration responseTime = finishTime - itRequest->second.StartTime;
-                tabletId = itRequest->second.TabletId;
-                TTabletState& tabletState = TabletStates[tabletId];
-                if (responseTime > tabletState.MaxResponseTime) {
-                    tabletState.MaxResponseTime = responseTime;
-                }
-                RequestsInFlight.erase(itRequest);
-            }
-            return tabletId;
-        }
-    };
 
     TTabletRequestsState TabletRequests;
 
@@ -1517,556 +1332,6 @@ public:
         }
     }
 
-    static TString GetNodeLocation(const TEvInterconnect::TNodeInfo& nodeInfo) {
-        return TStringBuilder() << nodeInfo.NodeId << '/' << nodeInfo.Host << ':' << nodeInfo.Port;
-    }
-
-    static void Check(TSelfCheckResult& context, const NKikimrWhiteboard::TSystemStateInfo::TPoolStats& poolStats) {
-        if (poolStats.name() == "System" || poolStats.name() == "IC" || poolStats.name() == "IO") {
-            if (poolStats.usage() >= 0.99) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Pool usage is over 99%", ETags::OverloadState);
-            } else if (poolStats.usage() >= 0.95) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Pool usage is over 95%", ETags::OverloadState);
-            } else {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-            }
-        } else {
-            if (poolStats.usage() >= 0.99) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Pool usage is over 99%", ETags::OverloadState);
-            } else {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-            }
-        }
-    }
-
-    Ydb::Monitoring::StatusFlag::Status FillSystemTablets(TDatabaseState& databaseState, TSelfCheckResult context) {
-        TString databaseId = context.Location.database().name();
-        for (auto& [tabletId, tablet] : TabletRequests.TabletStates) {
-            if (tablet.Database == databaseId) {
-                auto tabletIt = databaseState.MergedTabletState.find(std::make_pair(tabletId, 0));
-                if (tabletIt != databaseState.MergedTabletState.end()) {
-                    auto nodeId = tabletIt->second->GetNodeID();
-                    if (nodeId) {
-                        FillNodeInfo(nodeId, context.Location.mutable_node());
-                    }
-                }
-
-                context.Location.mutable_compute()->clear_tablet();
-                auto& protoTablet = *context.Location.mutable_compute()->mutable_tablet();
-                auto timeoutMs = Timeout.MilliSeconds();
-                auto orangeTimeout = timeoutMs / 2;
-                auto yellowTimeout = timeoutMs / 10;
-                if (tablet.IsUnresponsive || tablet.MaxResponseTime >= TDuration::MilliSeconds(yellowTimeout)) {
-                    if (tablet.Type != TTabletTypes::Unknown) {
-                        protoTablet.set_type(TTabletTypes::EType_Name(tablet.Type));
-                    }
-                    protoTablet.add_id(ToString(tabletId));
-                    if (tablet.IsUnresponsive) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet is unresponsive", ETags::SystemTabletState);
-                    } else if (tablet.MaxResponseTime >= TDuration::MilliSeconds(orangeTimeout)) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, TStringBuilder() << "System tablet response time is over " << orangeTimeout << "ms", ETags::SystemTabletState);
-                    } else if (tablet.MaxResponseTime >= TDuration::MilliSeconds(yellowTimeout)) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, TStringBuilder() << "System tablet response time is over " << yellowTimeout << "ms", ETags::SystemTabletState);
-                    }
-                }
-            }
-        }
-        return context.GetOverallStatus();
-    }
-
-    Ydb::Monitoring::StatusFlag::Status FillTablets(TDatabaseState& databaseState,
-                                                    TNodeId nodeId,
-                                                    google::protobuf::RepeatedPtrField<Ydb::Monitoring::ComputeTabletStatus>& parent,
-                                                    TSelfCheckResult& context) {
-        Ydb::Monitoring::StatusFlag::Status tabletsStatus = Ydb::Monitoring::StatusFlag::GREEN;
-        auto itNodeTabletState = databaseState.MergedNodeTabletState.find(nodeId);
-        if (itNodeTabletState != databaseState.MergedNodeTabletState.end()) {
-            TSelfCheckResult tabletsContext(&context);
-            for (const auto& count : itNodeTabletState->second.Count) {
-                if (count.Count > 0) {
-                    TSelfCheckResult tabletContext(&tabletsContext, "TABLET");
-                    auto& protoTablet = *tabletContext.Location.mutable_compute()->mutable_tablet();
-                    FillNodeInfo(nodeId, tabletContext.Location.mutable_node());
-                    protoTablet.set_type(TTabletTypes::EType_Name(count.Type));
-                    protoTablet.set_count(count.Count);
-                    if (!count.Identifiers.empty()) {
-                        for (const TString& id : count.Identifiers) {
-                            protoTablet.add_id(id);
-                        }
-                    }
-                    Ydb::Monitoring::ComputeTabletStatus& computeTabletStatus = *parent.Add();
-                    computeTabletStatus.set_type(NKikimrTabletBase::TTabletTypes::EType_Name(count.Type));
-                    computeTabletStatus.set_count(count.Count);
-                    for (const TString& id : count.Identifiers) {
-                        computeTabletStatus.add_id(id);
-                    }
-                    switch (count.State) {
-                        case TNodeTabletState::ETabletState::Good:
-                            computeTabletStatus.set_state("GOOD");
-                            tabletContext.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                            break;
-                        case TNodeTabletState::ETabletState::Stopped:
-                            computeTabletStatus.set_state("STOPPED");
-                            tabletContext.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                            break;
-                        case TNodeTabletState::ETabletState::RestartsTooOften:
-                            computeTabletStatus.set_state("RESTARTS_TOO_OFTEN");
-                            tabletContext.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Tablets are restarting too often", ETags::TabletState);
-                            break;
-                        case TNodeTabletState::ETabletState::Dead:
-                            computeTabletStatus.set_state("DEAD");
-                            if (count.Leader) {
-                                tabletContext.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Tablets are dead", ETags::TabletState);
-                            } else {
-                                tabletContext.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Followers are dead", ETags::TabletState);
-                            }
-                            break;
-                    }
-                    computeTabletStatus.set_overall(tabletContext.GetOverallStatus());
-                    tabletsStatus = MaxStatus(tabletsStatus, tabletContext.GetOverallStatus());
-                }
-            }
-        }
-        return tabletsStatus;
-    }
-
-    void FillNodeInfo(TNodeId nodeId, Ydb::Monitoring::LocationNode* node) {
-        const TEvInterconnect::TNodeInfo* nodeInfo = nullptr;
-        auto itNodeInfo = MergedNodeInfo.find(nodeId);
-        if (itNodeInfo != MergedNodeInfo.end()) {
-            nodeInfo = itNodeInfo->second;
-        }
-        TString id(ToString(nodeId));
-
-        node->set_id(nodeId);
-        if (nodeInfo) {
-            node->set_host(nodeInfo->Host);
-            node->set_port(nodeInfo->Port);
-        }
-    }
-
-    void FillComputeNodeStatus(TDatabaseState& databaseState, TNodeId nodeId, Ydb::Monitoring::ComputeNodeStatus& computeNodeStatus, TSelfCheckResult context) {
-        FillNodeInfo(nodeId, context.Location.mutable_compute()->mutable_node());
-
-        TSelfCheckResult rrContext(&context, "NODE_UPTIME");
-        if (databaseState.NodeRestartsPerPeriod[nodeId] >= HealthCheckConfig.GetThresholds().GetNodeRestartsOrange()) {
-            rrContext.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Node is restarting too often", ETags::Uptime);
-        } else if (databaseState.NodeRestartsPerPeriod[nodeId] >= HealthCheckConfig.GetThresholds().GetNodeRestartsYellow()) {
-            rrContext.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "The number of node restarts has increased", ETags::Uptime);
-        } else {
-            rrContext.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-        }
-
-        auto itNodeSystemState = MergedNodeSystemState.find(nodeId);
-        if (itNodeSystemState != MergedNodeSystemState.end()) {
-            const NKikimrWhiteboard::TSystemStateInfo& nodeSystemState(*itNodeSystemState->second);
-
-            for (const auto& poolStat : nodeSystemState.poolstats()) {
-                TSelfCheckResult poolContext(&context, "COMPUTE_POOL");
-                poolContext.Location.mutable_compute()->mutable_pool()->set_name(poolStat.name());
-                Check(poolContext, poolStat);
-                Ydb::Monitoring::ThreadPoolStatus& threadPoolStatus = *computeNodeStatus.add_pools();
-                threadPoolStatus.set_name(poolStat.name());
-                threadPoolStatus.set_usage(poolStat.usage());
-                threadPoolStatus.set_overall(poolContext.GetOverallStatus());
-            }
-
-            if (nodeSystemState.loadaverage_size() > 0 && nodeSystemState.numberofcpus() > 0) {
-                TSelfCheckResult laContext(&context, "LOAD_AVERAGE");
-                Ydb::Monitoring::LoadAverageStatus& loadAverageStatus = *computeNodeStatus.mutable_load();
-                loadAverageStatus.set_load(nodeSystemState.loadaverage(0));
-                loadAverageStatus.set_cores(nodeSystemState.numberofcpus());
-                if (loadAverageStatus.load() > loadAverageStatus.cores()) {
-                    laContext.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "LoadAverage above 100%", ETags::OverloadState);
-                } else {
-                    laContext.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                }
-                loadAverageStatus.set_overall(laContext.GetOverallStatus());
-            }
-
-            if (nodeSystemState.HasMaxClockSkewPeerId()) {
-                TNodeId peerId = nodeSystemState.GetMaxClockSkewPeerId();
-                long timeDifferenceUs = nodeSystemState.GetMaxClockSkewWithPeerUs();
-                TDuration timeDifferenceDuration = TDuration::MicroSeconds(abs(timeDifferenceUs));
-                Ydb::Monitoring::StatusFlag::Status status;
-                if (timeDifferenceDuration > TDuration::MicroSeconds(HealthCheckConfig.GetThresholds().GetNodesTimeDifferenceOrange())) {
-                    status = Ydb::Monitoring::StatusFlag::ORANGE;
-                } else if (timeDifferenceDuration > TDuration::MicroSeconds(HealthCheckConfig.GetThresholds().GetNodesTimeDifferenceYellow())) {
-                    status = Ydb::Monitoring::StatusFlag::YELLOW;
-                } else {
-                    status = Ydb::Monitoring::StatusFlag::GREEN;
-                }
-
-                if (databaseState.MaxTimeDifferenceNodeId == nodeId) {
-                    TSelfCheckResult tdContext(&context, "NODES_TIME_DIFFERENCE");
-                    if (status == Ydb::Monitoring::StatusFlag::GREEN) {
-                        tdContext.ReportStatus(status);
-                    } else {
-                        tdContext.ReportStatus(status, TStringBuilder() << "Node is  "
-                                                                        << timeDifferenceDuration.MilliSeconds() << " ms "
-                                                                        << (timeDifferenceUs > 0 ? "behind " : "ahead of ")
-                                                                        << "peer [" << peerId << "]", ETags::SyncState);
-                    }
-                }
-            }
-        } else {
-            // context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-            //                      TStringBuilder() << "Compute node is not available",
-             //                      ETags::NodeState);
-        }
-        computeNodeStatus.set_id(ToString(nodeId));
-        computeNodeStatus.set_overall(context.GetOverallStatus());
-    }
-
-    void FillComputeDatabaseStatus(TDatabaseState& databaseState, Ydb::Monitoring::ComputeStatus& computeStatus, TSelfCheckResult context) {
-        auto itDescribe = DescribeByPath.find(databaseState.Path);
-        if (itDescribe != DescribeByPath.end() && itDescribe->second.IsOk()) {
-            const auto& domain(itDescribe->second->GetRecord().GetPathDescription().GetDomainDescription());
-            if (domain.GetPathsLimit() > 0) {
-                float usage = (float)domain.GetPathsInside() / domain.GetPathsLimit();
-                computeStatus.set_paths_quota_usage(usage);
-                if (static_cast<i64>(domain.GetPathsLimit()) - static_cast<i64>(domain.GetPathsInside()) <= 1) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Paths quota exhausted", ETags::QuotaUsage);
-                } else if (usage >= 0.99) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Paths quota usage is over 99%", ETags::QuotaUsage);
-                } else if (usage >= 0.90) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Paths quota usage is over 90%", ETags::QuotaUsage);
-                } else {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                }
-            }
-            if (domain.GetShardsLimit() > 0) {
-                float usage = (float)domain.GetShardsInside() / domain.GetShardsLimit();
-                computeStatus.set_shards_quota_usage(usage);
-                if (static_cast<i64>(domain.GetShardsLimit()) - static_cast<i64>(domain.GetShardsInside()) <= 1) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Shards quota exhausted", ETags::QuotaUsage);
-                } else if (usage >= 0.99) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Shards quota usage is over 99%", ETags::QuotaUsage);
-                } else if (usage >= 0.90) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Shards quota usage is over 90%", ETags::QuotaUsage);
-                } else {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                }
-            }
-        }
-    }
-
-    void FillCompute(TDatabaseState& databaseState, Ydb::Monitoring::ComputeStatus& computeStatus, TSelfCheckResult context) {
-        TVector<TNodeId>* computeNodeIds = &databaseState.ComputeNodeIds;
-        if (databaseState.ResourcePathId
-            && databaseState.ServerlessComputeResourcesMode != NKikimrSubDomains::EServerlessComputeResourcesModeExclusive)
-        {
-            auto itDatabase = FilterDomainKey.find(TSubDomainKey(databaseState.ResourcePathId.OwnerId, databaseState.ResourcePathId.LocalPathId));
-            if (itDatabase != FilterDomainKey.end()) {
-                const TString& sharedDatabaseName = itDatabase->second;
-                TDatabaseState& sharedDatabase = DatabaseState[sharedDatabaseName];
-                computeNodeIds = &sharedDatabase.ComputeNodeIds;
-            }
-        }
-        std::sort(computeNodeIds->begin(), computeNodeIds->end());
-        computeNodeIds->erase(std::unique(computeNodeIds->begin(), computeNodeIds->end()), computeNodeIds->end());
-        if (computeNodeIds->empty()) {
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "There are no compute nodes", ETags::ComputeState);
-        } else {
-            long maxTimeDifferenceUs = 0;
-            for (TNodeId nodeId : *computeNodeIds) {
-                auto itNodeSystemState = MergedNodeSystemState.find(nodeId);
-                if (itNodeSystemState != MergedNodeSystemState.end()) {
-                    if (std::count(computeNodeIds->begin(), computeNodeIds->end(), itNodeSystemState->second->GetMaxClockSkewPeerId()) > 0
-                            && abs(itNodeSystemState->second->GetMaxClockSkewWithPeerUs()) > maxTimeDifferenceUs) {
-                        maxTimeDifferenceUs = abs(itNodeSystemState->second->GetMaxClockSkewWithPeerUs());
-                        databaseState.MaxTimeDifferenceNodeId = nodeId;
-                    }
-                }
-            }
-            for (TNodeId nodeId : *computeNodeIds) {
-                auto& computeNode = *computeStatus.add_nodes();
-                FillComputeNodeStatus(databaseState, nodeId, computeNode, {&context, "COMPUTE_NODE"});
-            }
-        }
-        Ydb::Monitoring::StatusFlag::Status systemStatus = FillSystemTablets(databaseState, {&context, "SYSTEM_TABLET"});
-        if (systemStatus != Ydb::Monitoring::StatusFlag::GREEN && systemStatus != Ydb::Monitoring::StatusFlag::GREY) {
-            context.ReportStatus(systemStatus, "Compute has issues with system tablets", ETags::ComputeState, {ETags::SystemTabletState});
-        }
-        FillComputeDatabaseStatus(databaseState, computeStatus, {&context, "COMPUTE_QUOTA"});
-        context.ReportWithMaxChildStatus("Some nodes are restarting too often", ETags::ComputeState, {ETags::Uptime});
-        context.ReportWithMaxChildStatus("Compute is overloaded", ETags::ComputeState, {ETags::OverloadState});
-        context.ReportWithMaxChildStatus("Compute quota usage", ETags::ComputeState, {ETags::QuotaUsage});
-        context.ReportWithMaxChildStatus("Database has time difference between nodes", ETags::ComputeState, {ETags::SyncState});
-        Ydb::Monitoring::StatusFlag::Status tabletsStatus = Ydb::Monitoring::StatusFlag::GREEN;
-        computeNodeIds->push_back(0); // for tablets without node
-        for (TNodeId nodeId : *computeNodeIds) {
-            tabletsStatus = MaxStatus(tabletsStatus, FillTablets(databaseState, nodeId, *computeStatus.mutable_tablets(), context));
-        }
-        if (tabletsStatus != Ydb::Monitoring::StatusFlag::GREEN) {
-            context.ReportStatus(tabletsStatus, "Compute has issues with tablets", ETags::ComputeState, {ETags::TabletState});
-        }
-        if (ReturnHints) {
-            auto schemeShardId = databaseState.SchemeShardId;
-            if (schemeShardId) {
-                for (const auto& [path, hint] : OverloadedShardHints) {
-                    if (hint.SchemeShardId == schemeShardId) {
-                        TSelfCheckResult hintContext(&context, "HINT-OVERLOADED-SHARD");
-                        //hintContext.Location.mutable_compute()->mutable_tablet()->set_type(NKikimrTabletBase::TTabletTypes::EType_Name(NKikimrTabletBase::TTabletTypes::DataShard));
-                        TStringBuilder tabletId;
-                        tabletId << hint.TabletId;
-                        if (hint.FollowerId) {
-                            tabletId << '-' << hint.FollowerId;
-                        }
-                        hintContext.Location.mutable_compute()->mutable_tablet()->add_id(tabletId);
-                        hintContext.Location.mutable_compute()->mutable_schema()->set_type("table");
-                        hintContext.Location.mutable_compute()->mutable_schema()->set_path(path);
-                        hintContext.ReportStatus(Ydb::Monitoring::StatusFlag::UNSPECIFIED, hint.Message);
-                    }
-                }
-            }
-        }
-        computeStatus.set_overall(context.GetOverallStatus());
-    }
-
-    static TString GetVSlotId(const NKikimrSysView::TVSlotKey& vSlotKey) {
-        return TStringBuilder()
-                << vSlotKey.GetNodeId() << '-'
-                << vSlotKey.GetPDiskId() << '-'
-                << vSlotKey.GetVSlotId();
-    }
-
-    static TString GetVDiskId(const NKikimrBlobStorage::TVDiskID& protoVDiskId) {
-        return TStringBuilder()
-                << protoVDiskId.groupid() << '-'
-                << protoVDiskId.groupgeneration() << '-'
-                << protoVDiskId.ring() << '-'
-                << protoVDiskId.domain() << '-'
-                << protoVDiskId.vdisk();
-    }
-
-    static TString GetVDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& protoVSlot) {
-        return TStringBuilder()
-                << protoVSlot.groupid() << '-'
-                << protoVSlot.groupgeneration() << '-'
-                << protoVSlot.failrealmidx() << '-'
-                << protoVSlot.faildomainidx() << '-'
-                << protoVSlot.vdiskidx();
-    }
-
-    static TString GetVDiskId(const NKikimrBlobStorage::TNodeWardenServiceSet_TVDisk& protoVDiskId) {
-        return GetVDiskId(protoVDiskId.vdiskid());
-    }
-
-    static TString GetVDiskId(const NKikimrWhiteboard::TVDiskStateInfo vDiskInfo) {
-        return GetVDiskId(vDiskInfo.vdiskid());
-    }
-
-    static TString GetVDiskId(const NKikimrSysView::TVSlotInfo& vSlot) {
-        return TStringBuilder()
-               << vSlot.GetGroupId() << '-'
-               << vSlot.GetGroupGeneration() << '-'
-               << vSlot.GetFailRealm() << '-'
-               << vSlot.GetFailDomain() << '-'
-               << vSlot.GetVDisk();
-    }
-
-    static TString GetPDiskId(const NKikimrWhiteboard::TVDiskStateInfo vDiskInfo) {
-        return TStringBuilder() << vDiskInfo.nodeid() << "-" << vDiskInfo.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrWhiteboard::TPDiskStateInfo pDiskInfo) {
-        return TStringBuilder() << pDiskInfo.nodeid() << "-" << pDiskInfo.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TPDisk& pDisk) {
-        return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& vSlot) {
-        return TStringBuilder() << vSlot.vslotid().nodeid() << "-" << vSlot.vslotid().pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrBlobStorage::TNodeWardenServiceSet_TPDisk& pDisk) {
-        return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrSysView::TVSlotKey& vSlotKey) {
-        return TStringBuilder() << vSlotKey.GetNodeId() << "-" << vSlotKey.GetPDiskId();
-    }
-
-    static TString GetPDiskId(const NKikimrSysView::TPDiskKey& pDiskKey) {
-        return TStringBuilder() << pDiskKey.GetNodeId() << "-" << pDiskKey.GetPDiskId();
-    }
-
-    void FillPDiskStatus(const TString& pDiskId, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckResult context) {
-        context.Location.clear_database(); // PDisks are shared between databases
-        context.Location.mutable_storage()->mutable_pool()->clear_name(); // PDisks are shared between pools
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // PDisks are shared between groups
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->clear_id(); // PDisks are shared between vdisks
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->Clear();
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_pdisk();
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_id(pDiskId);
-        storagePDiskStatus.set_id(pDiskId);
-
-        auto itPDisk = PDisksMap.find(pDiskId);
-        if (itPDisk == PDisksMap.end()) { // this report, in theory, can't happen because there was pdisk mention in bsc vslot info. this pdisk info have to exists in bsc too
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet BSC didn't provide expected pdisk information", ETags::PDiskState);
-            storagePDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        const auto& pDisk = itPDisk->second->GetInfo();
-
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_path(pDisk.GetPath());
-        const auto& statusString = pDisk.GetStatusV2();
-        const auto *descriptor = NKikimrBlobStorage::EDriveStatus_descriptor();
-        auto status = descriptor->FindValueByName(statusString);
-        if (!status) {
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                 TStringBuilder() << "Unknown PDisk state: " << statusString,
-                                 ETags::PDiskState);
-        }
-        switch (status->number()) {
-            case NKikimrBlobStorage::ACTIVE:
-            case NKikimrBlobStorage::INACTIVE: {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                break;
-            }
-            case NKikimrBlobStorage::FAULTY: {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW,
-                                     TStringBuilder() << "PDisk state is " << statusString,
-                                     ETags::PDiskState);
-                break;
-            }
-            case NKikimrBlobStorage::BROKEN:
-            case NKikimrBlobStorage::TO_BE_REMOVED: {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                     TStringBuilder() << "PDisk state is " << statusString,
-                                     ETags::PDiskState);
-                break;
-            }
-        }
-
-        if (pDisk.GetAvailableSize() != 0 && pDisk.GetTotalSize() != 0) { // do not replace it with Has()
-            double avail = (double)pDisk.GetAvailableSize() / pDisk.GetTotalSize();
-            if (avail < 0.06) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Available size is less than 6%", ETags::PDiskSpace);
-            } else if (avail < 0.09) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Available size is less than 9%", ETags::PDiskSpace);
-            } else if (avail < 0.12) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Available size is less than 12%", ETags::PDiskSpace);
-            }
-        }
-
-        // do not propagate RED status to vdisk - so that vdisk is not considered down when computing group status
-        context.OverallStatus = MinStatus(context.OverallStatus, Ydb::Monitoring::StatusFlag::ORANGE);
-        storagePDiskStatus.set_overall(context.GetOverallStatus());
-    }
-
-    static Ydb::Monitoring::StatusFlag::Status GetFlagFromBSPDiskSpaceColor(NKikimrBlobStorage::TPDiskSpaceColor::E flag) {
-        switch (flag) {
-            case NKikimrBlobStorage::TPDiskSpaceColor::GREEN:
-            case NKikimrBlobStorage::TPDiskSpaceColor::CYAN:
-                return Ydb::Monitoring::StatusFlag::GREEN;
-            case NKikimrBlobStorage::TPDiskSpaceColor::LIGHT_YELLOW:
-            case NKikimrBlobStorage::TPDiskSpaceColor::YELLOW:
-                return Ydb::Monitoring::StatusFlag::YELLOW;
-            case NKikimrBlobStorage::TPDiskSpaceColor::LIGHT_ORANGE:
-            case NKikimrBlobStorage::TPDiskSpaceColor::PRE_ORANGE:
-            case NKikimrBlobStorage::TPDiskSpaceColor::ORANGE:
-                return Ydb::Monitoring::StatusFlag::ORANGE;
-            case NKikimrBlobStorage::TPDiskSpaceColor::RED:
-                return Ydb::Monitoring::StatusFlag::RED;
-            default:
-                return Ydb::Monitoring::StatusFlag::UNSPECIFIED;
-        }
-    }
-
-    static Ydb::Monitoring::StatusFlag::Status GetFlagFromWhiteboardFlag(NKikimrWhiteboard::EFlag flag) {
-        switch (flag) {
-            case NKikimrWhiteboard::EFlag::Green:
-                return Ydb::Monitoring::StatusFlag::GREEN;
-            case NKikimrWhiteboard::EFlag::Yellow:
-                return Ydb::Monitoring::StatusFlag::YELLOW;
-            case NKikimrWhiteboard::EFlag::Orange:
-                return Ydb::Monitoring::StatusFlag::ORANGE;
-            case NKikimrWhiteboard::EFlag::Red:
-                return Ydb::Monitoring::StatusFlag::RED;
-            default:
-                return Ydb::Monitoring::StatusFlag::UNSPECIFIED;
-        }
-    }
-
-    void FillVDiskStatus(const NKikimrSysView::TVSlotEntry* vSlot, Ydb::Monitoring::StorageVDiskStatus& storageVDiskStatus, TSelfCheckResult context) {
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id()->Clear();
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // you can see VDisks Group Id in vSlotId field
-
-        TNodeId nodeId = vSlot->GetKey().GetNodeId();
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_id(GetVDiskId(vSlot->GetInfo()));
-        context.Location.mutable_storage()->mutable_node()->set_id(nodeId);
-
-        auto itNodeInfo = MergedNodeInfo.find(nodeId);
-        if (itNodeInfo != MergedNodeInfo.end()) {
-            context.Location.mutable_storage()->mutable_node()->set_host(itNodeInfo->second->Host);
-            context.Location.mutable_storage()->mutable_node()->set_port(itNodeInfo->second->Port);
-        } else {
-            context.Location.mutable_storage()->mutable_node()->clear_host();
-            context.Location.mutable_storage()->mutable_node()->clear_port();
-        }
-
-        storageVDiskStatus.set_id(GetVSlotId(vSlot->GetKey()));
-
-        const auto& vSlotInfo = vSlot->GetInfo();
-
-        if (!vSlotInfo.HasStatusV2()) {
-            // this should mean that BSC recently restarted and does not have accurate data yet - we should not report to avoid false positives
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-            storageVDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        const auto *descriptor = NKikimrBlobStorage::EVDiskStatus_descriptor();
-        auto status = descriptor->FindValueByName(vSlot->GetInfo().GetStatusV2());
-        if (!status) { // this case is not expected because becouse bsc assignes status according EVDiskStatus enum
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet BSC didn't provide known status", ETags::VDiskState);
-            storageVDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        if (vSlot->GetKey().HasPDiskId()) {
-            TString pDiskId = GetPDiskId(vSlot->GetKey());
-            FillPDiskStatus(pDiskId, *storageVDiskStatus.mutable_pdisk(), {&context, "PDISK"});
-        }
-
-        if (status->number() == NKikimrBlobStorage::ERROR) {
-            // the disk is not operational at all
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "VDisk is not available", ETags::VDiskState,{ETags::PDiskState});
-            storageVDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        if (vSlotInfo.HasIsThrottling() && vSlotInfo.GetIsThrottling()) {
-            // throttling is active
-            auto message = TStringBuilder() << "VDisk is being throttled, rate "
-                << vSlotInfo.GetThrottlingRate() << " per mille";
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, message, ETags::VDiskState);
-            storageVDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        switch (status->number()) {
-            case NKikimrBlobStorage::REPLICATING: { // the disk accepts queries, but not all the data was replicated
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, TStringBuilder() << "Replication in progress", ETags::VDiskState);
-                storageVDiskStatus.set_overall(context.GetOverallStatus());
-                return;
-            }
-            case NKikimrBlobStorage::INIT_PENDING:
-            case NKikimrBlobStorage::READY: { // the disk is fully operational and does not affect group fault tolerance
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-            }
-            default:
-                break;
-        }
-
-        storageVDiskStatus.set_overall(context.GetOverallStatus());
-    }
-
     void Handle(TEvWhiteboard::TEvVDiskStateResponse::TPtr& ev) {
         TNodeId nodeId = ev.Get()->Cookie;
         auto& nodeVDiskState(NodeVDiskState[nodeId]);
@@ -2112,807 +1377,6 @@ public:
         RequestDone("TEvBSGroupStateResponse");
     }
 
-    void FillPDiskStatusWithWhiteboard(const TString& pDiskId, const NKikimrWhiteboard::TPDiskStateInfo& pDiskInfo, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckResult context) {
-        context.Location.clear_database(); // PDisks are shared between databases
-        if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->empty()) {
-            context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_pdisk();
-        }
-        context.Location.mutable_storage()->mutable_pool()->clear_name(); // PDisks are shared between pools
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // PDisks are shared between groups
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->clear_id(); // PDisks are shared between vdisks
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_id(pDiskId);
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_path(pDiskInfo.path());
-        storagePDiskStatus.set_id(pDiskId);
-
-        if (pDiskInfo.HasState()) {
-            switch (pDiskInfo.GetState()) {
-                case NKikimrBlobStorage::TPDiskState::Normal:
-                case NKikimrBlobStorage::TPDiskState::Stopped:
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                    break;
-                case NKikimrBlobStorage::TPDiskState::Initial:
-                case NKikimrBlobStorage::TPDiskState::InitialFormatRead:
-                case NKikimrBlobStorage::TPDiskState::InitialSysLogRead:
-                case NKikimrBlobStorage::TPDiskState::InitialCommonLogRead:
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW,
-                                         TStringBuilder() << "PDisk state is " << NKikimrBlobStorage::TPDiskState::E_Name(pDiskInfo.GetState()),
-                                         ETags::PDiskState);
-                    break;
-                case NKikimrBlobStorage::TPDiskState::InitialFormatReadError:
-                case NKikimrBlobStorage::TPDiskState::InitialSysLogReadError:
-                case NKikimrBlobStorage::TPDiskState::InitialSysLogParseError:
-                case NKikimrBlobStorage::TPDiskState::InitialCommonLogReadError:
-                case NKikimrBlobStorage::TPDiskState::InitialCommonLogParseError:
-                case NKikimrBlobStorage::TPDiskState::CommonLoggerInitError:
-                case NKikimrBlobStorage::TPDiskState::OpenFileError:
-                case NKikimrBlobStorage::TPDiskState::ChunkQuotaError:
-                case NKikimrBlobStorage::TPDiskState::DeviceIoError:
-                case NKikimrBlobStorage::TPDiskState::Missing:
-                case NKikimrBlobStorage::TPDiskState::Timeout:
-                case NKikimrBlobStorage::TPDiskState::NodeDisconnected:
-                case NKikimrBlobStorage::TPDiskState::Unknown:
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                         TStringBuilder() << "PDisk state is " << NKikimrBlobStorage::TPDiskState::E_Name(pDiskInfo.GetState()),
-                                         ETags::PDiskState);
-                    break;
-                case NKikimrBlobStorage::TPDiskState::Reserved15:
-                case NKikimrBlobStorage::TPDiskState::Reserved16:
-                case NKikimrBlobStorage::TPDiskState::Reserved17:
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Unknown PDisk state");
-                    break;
-            }
-
-            //if (pDiskInfo.HasAvailableSize() && pDiskInfo.GetTotalSize() != 0) {
-            if (pDiskInfo.GetAvailableSize() != 0 && pDiskInfo.GetTotalSize() != 0) { // hotfix until KIKIMR-12659
-                double avail = (double)pDiskInfo.GetAvailableSize() / pDiskInfo.GetTotalSize();
-                if (avail < 0.06) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Available size is less than 6%", ETags::PDiskSpace);
-                } else if (avail < 0.09) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Available size is less than 9%", ETags::PDiskSpace);
-                } else if (avail < 0.12) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Available size is less than 12%", ETags::PDiskSpace);
-                }
-            }
-        } else {
-            if (UnavailableStorageNodes.count(pDiskInfo.nodeid()) != 0) {
-                TSelfCheckResult nodeContext(&context, "STORAGE_NODE");
-                nodeContext.Location.mutable_storage()->clear_pool();
-                nodeContext.Location.mutable_storage()->mutable_node()->set_id(pDiskInfo.nodeid());
-                const TEvInterconnect::TNodeInfo* nodeInfo = nullptr;
-                auto itNodeInfo = MergedNodeInfo.find(pDiskInfo.nodeid());
-                if (itNodeInfo != MergedNodeInfo.end()) {
-                    nodeInfo = itNodeInfo->second;
-                }
-                if (nodeInfo) {
-                    nodeContext.Location.mutable_storage()->mutable_node()->set_host(nodeInfo->Host);
-                    nodeContext.Location.mutable_storage()->mutable_node()->set_port(nodeInfo->Port);
-                }
-                nodeContext.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                         TStringBuilder() << "Storage node is not available",
-                                         ETags::NodeState);
-            }
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                 TStringBuilder() << "PDisk is not available",
-                                 ETags::PDiskState,
-                                 {ETags::NodeState});
-        }
-
-        storagePDiskStatus.set_overall(context.GetOverallStatus());
-    }
-
-    void FillVDiskStatusWithWhiteboard(const TString& vDiskId, const NKikimrWhiteboard::TVDiskStateInfo& vDiskInfo, Ydb::Monitoring::StorageVDiskStatus& storageVDiskStatus, TSelfCheckResult context) {
-        if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id()->empty()) {
-            context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_id();
-        }
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->set_id(0, vDiskId);
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // you can see VDisks Group Id in vDiskId field
-        storageVDiskStatus.set_id(vDiskId);
-        TString pDiskId = GetPDiskId(vDiskInfo);
-        auto itPDisk = MergedPDiskState.find(pDiskId);
-        if (itPDisk != MergedPDiskState.end()) {
-            FillPDiskStatusWithWhiteboard(pDiskId, *itPDisk->second, *storageVDiskStatus.mutable_pdisk(), {&context, "PDISK"});
-        }
-
-        if (!vDiskInfo.HasVDiskState()) {
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                 TStringBuilder() << "VDisk is not available",
-                                 ETags::VDiskState,
-                                 {ETags::PDiskState});
-            storageVDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        switch (vDiskInfo.GetVDiskState()) {
-            case NKikimrWhiteboard::EVDiskState::OK:
-            case NKikimrWhiteboard::EVDiskState::Initial:
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                break;
-            case NKikimrWhiteboard::EVDiskState::SyncGuidRecovery:
-                context.IssueRecords.clear();
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW,
-                                     TStringBuilder() << "VDisk state is " << NKikimrWhiteboard::EVDiskState_Name(vDiskInfo.GetVDiskState()),
-                                     ETags::VDiskState);
-                storageVDiskStatus.set_overall(context.GetOverallStatus());
-                return;
-            case NKikimrWhiteboard::EVDiskState::LocalRecoveryError:
-            case NKikimrWhiteboard::EVDiskState::SyncGuidRecoveryError:
-            case NKikimrWhiteboard::EVDiskState::PDiskError:
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
-                                     TStringBuilder() << "VDisk state is " << NKikimrWhiteboard::EVDiskState_Name(vDiskInfo.GetVDiskState()),
-                                     ETags::VDiskState,
-                                     {ETags::PDiskState});
-                storageVDiskStatus.set_overall(context.GetOverallStatus());
-                return;
-        }
-
-        if (!vDiskInfo.GetReplicated()) {
-            context.IssueRecords.clear();
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, "Replication in progress", ETags::VDiskState);
-            storageVDiskStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        if (vDiskInfo.HasDiskSpace()) {
-            switch(vDiskInfo.GetDiskSpace()) {
-                case NKikimrWhiteboard::EFlag::Green:
-                    if (context.IssueRecords.size() == 0) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                    } else {
-                        context.ReportStatus(context.IssueRecords.begin()->IssueLog.status(),
-                                            TStringBuilder() << "VDisk is degraded",
-                                            ETags::VDiskState,
-                                            {ETags::PDiskSpace});
-                    }
-                    break;
-                case NKikimrWhiteboard::EFlag::Red:
-                    context.ReportStatus(GetFlagFromWhiteboardFlag(vDiskInfo.GetDiskSpace()),
-                                         TStringBuilder() << "DiskSpace is " << NKikimrWhiteboard::EFlag_Name(vDiskInfo.GetDiskSpace()),
-                                         ETags::VDiskState,
-                                         {ETags::PDiskSpace});
-                    break;
-                default:
-                    context.ReportStatus(GetFlagFromWhiteboardFlag(vDiskInfo.GetDiskSpace()),
-                                         TStringBuilder() << "DiskSpace is " << NKikimrWhiteboard::EFlag_Name(vDiskInfo.GetDiskSpace()),
-                                         ETags::VDiskSpace,
-                                         {ETags::PDiskSpace});
-                    break;
-            }
-        }
-
-        storageVDiskStatus.set_overall(context.GetOverallStatus());
-    }
-
-    class TGroupChecker {
-        TString ErasureSpecies;
-        bool LayoutCorrect;
-        int FailedDisks = 0;
-        std::array<int, Ydb::Monitoring::StatusFlag::Status_ARRAYSIZE> DisksColors = {};
-        TStackVec<std::pair<ui32, int>> FailedRealms;
-
-        void IncrementFor(ui32 realm) {
-            auto itRealm = FindIf(FailedRealms, [realm](const std::pair<ui32, int>& p) -> bool {
-                return p.first == realm;
-            });
-            if (itRealm == FailedRealms.end()) {
-                itRealm = FailedRealms.insert(FailedRealms.end(), { realm, 1 });
-            } else {
-                itRealm->second++;
-            }
-        }
-
-    public:
-        TGroupChecker(const TString& erasure, const bool layoutCorrect = true)
-            : ErasureSpecies(erasure)
-            , LayoutCorrect(layoutCorrect)
-        {}
-
-        void AddVDiskStatus(Ydb::Monitoring::StatusFlag::Status status, ui32 realm) {
-            ++DisksColors[status];
-            switch (status) {
-                case Ydb::Monitoring::StatusFlag::BLUE: // disk is good, but not available
-                // No yellow or orange status here - this is intentional - they are used when a disk is running out of space, but is currently available
-                case Ydb::Monitoring::StatusFlag::RED: // disk is bad, probably not available
-                case Ydb::Monitoring::StatusFlag::GREY: // the status is absent, the disk is not available
-                    IncrementFor(realm);
-                    ++FailedDisks;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        void ReportStatus(TSelfCheckResult& context) const {
-            context.OverallStatus = Ydb::Monitoring::StatusFlag::GREEN;
-            if (!LayoutCorrect) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Group layout is incorrect", ETags::GroupState);
-            }
-            if (ErasureSpecies == NONE) {
-                if (FailedDisks > 0) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Group failed", ETags::GroupState, {ETags::VDiskState});
-                }
-            } else if (ErasureSpecies == BLOCK_4_2) {
-                if (FailedDisks > 2) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Group failed", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedDisks > 1) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Group has no redundancy", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedDisks > 0) {
-                    if (DisksColors[Ydb::Monitoring::StatusFlag::BLUE] == FailedDisks) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    } else {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    }
-                }
-            } else if (ErasureSpecies == MIRROR_3_DC) {
-                if (FailedRealms.size() > 2 || (FailedRealms.size() == 2 && FailedRealms[0].second > 1 && FailedRealms[1].second > 1)) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Group failed", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedRealms.size() == 2) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Group has no redundancy", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedDisks > 0) {
-                    if (DisksColors[Ydb::Monitoring::StatusFlag::BLUE] == FailedDisks) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    } else {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    }
-                }
-            }
-        }
-    };
-
-    void FillGroupStatusWithWhiteboard(TGroupId groupId, const NKikimrWhiteboard::TBSGroupStateInfo& groupInfo, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckResult context) {
-        if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_id()->empty()) {
-            context.Location.mutable_storage()->mutable_pool()->mutable_group()->add_id();
-        }
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->set_id(0, ToString(groupId));
-        storageGroupStatus.set_id(ToString(groupId));
-        TGroupChecker checker(groupInfo.erasurespecies());
-        for (const auto& protoVDiskId : groupInfo.vdiskids()) {
-            TString vDiskId = GetVDiskId(protoVDiskId);
-            auto itVDisk = MergedVDiskState.find(vDiskId);
-            const TEvInterconnect::TNodeInfo* nodeInfo = nullptr;
-            if (itVDisk != MergedVDiskState.end()) {
-                TNodeId nodeId = itVDisk->second->nodeid();
-                auto itNodeInfo = MergedNodeInfo.find(nodeId);
-                if (itNodeInfo != MergedNodeInfo.end()) {
-                    nodeInfo = itNodeInfo->second;
-                }
-                context.Location.mutable_storage()->mutable_node()->set_id(nodeId);
-            } else {
-                context.Location.mutable_storage()->mutable_node()->clear_id();
-            }
-            if (nodeInfo) {
-                context.Location.mutable_storage()->mutable_node()->set_host(nodeInfo->Host);
-                context.Location.mutable_storage()->mutable_node()->set_port(nodeInfo->Port);
-            } else {
-                context.Location.mutable_storage()->mutable_node()->clear_host();
-                context.Location.mutable_storage()->mutable_node()->clear_port();
-            }
-            Ydb::Monitoring::StorageVDiskStatus& vDiskStatus = *storageGroupStatus.add_vdisks();
-            FillVDiskStatusWithWhiteboard(vDiskId, itVDisk != MergedVDiskState.end() ? *itVDisk->second : NKikimrWhiteboard::TVDiskStateInfo(), vDiskStatus, {&context, "VDISK"});
-            checker.AddVDiskStatus(vDiskStatus.overall(), protoVDiskId.ring());
-        }
-
-        context.Location.mutable_storage()->clear_node(); // group doesn't have node
-        context.OverallStatus = MinStatus(context.OverallStatus, Ydb::Monitoring::StatusFlag::YELLOW);
-        checker.ReportStatus(context);
-
-        BLOG_D("Group " << groupId << " has status " << context.GetOverallStatus());
-        storageGroupStatus.set_overall(context.GetOverallStatus());
-    }
-
-    static const inline TString NONE = "none";
-    static const inline TString BLOCK_4_2 = "block-4-2";
-    static const inline TString MIRROR_3_DC = "mirror-3-dc";
-    static const int MERGING_IGNORE_SIZE = 4;
-
-    struct TMergeIssuesContext {
-        std::unordered_map<ETags, TList<TSelfCheckResult::TIssueRecord>> recordsMap;
-        std::unordered_set<TString> removeIssuesIds;
-
-        TMergeIssuesContext(TList<TSelfCheckResult::TIssueRecord>& records) {
-            for (auto it = records.begin(); it != records.end(); ) {
-                auto move = it++;
-                recordsMap[move->Tag].splice(recordsMap[move->Tag].end(), records, move);
-            }
-        }
-
-        void RemoveUnlinkIssues(TList<TSelfCheckResult::TIssueRecord>& records) {
-            bool isRemovingIssuesIteration = true;
-            while (isRemovingIssuesIteration) {
-                isRemovingIssuesIteration = false;
-
-                std::unordered_set<TString> necessaryIssuesIds;
-                for (auto it = records.begin(); it != records.end(); it++) {
-                    auto reasons = it->IssueLog.reason();
-                    for (auto reasonIt = reasons.begin(); reasonIt != reasons.end(); reasonIt++) {
-                        necessaryIssuesIds.insert(*reasonIt);
-                    }
-                }
-
-                for (auto it = records.begin(); it != records.end(); ) {
-                    if (!necessaryIssuesIds.contains(it->IssueLog.id()) && removeIssuesIds.contains(it->IssueLog.id())) {
-                        auto reasons = it->IssueLog.reason();
-                        for (auto reasonIt = reasons.begin(); reasonIt != reasons.end(); reasonIt++) {
-                            removeIssuesIds.insert(*reasonIt);
-                        }
-                        isRemovingIssuesIteration = true;
-                        it = records.erase(it);
-                    } else {
-                        it++;
-                    }
-                }
-            }
-
-            {
-                std::unordered_set<TString> issueIds;
-                for (auto it = records.begin(); it != records.end(); it++) {
-                    issueIds.insert(it->IssueLog.id());
-                }
-
-                for (auto it = records.begin(); it != records.end(); it++) {
-                    auto reasons = it->IssueLog.mutable_reason();
-                    for (auto reasonIt = reasons->begin(); reasonIt != reasons->end(); ) {
-                        if (!issueIds.contains(*reasonIt)) {
-                            reasonIt = reasons->erase(reasonIt);
-                        } else {
-                            reasonIt++;
-                        }
-                    }
-                }
-            }
-        }
-
-        void RenameMergingIssues(TList<TSelfCheckResult::TIssueRecord>& records) {
-            for (auto it = records.begin(); it != records.end(); it++) {
-                if (it->IssueLog.count() > 0) {
-                    TString message = it->IssueLog.message();
-                    switch (it->Tag) {
-                        case ETags::GroupState: {
-                            message = std::regex_replace(message.c_str(), std::regex("^Group has "), "Groups have ");
-                            message = std::regex_replace(message.c_str(), std::regex("^Group is "), "Groups are ");
-                            message = std::regex_replace(message.c_str(), std::regex("^Group "), "Groups ");
-                            break;
-                        }
-                        case ETags::VDiskState: {
-                            message = std::regex_replace(message.c_str(), std::regex("^VDisk has "), "VDisks have ");
-                            message = std::regex_replace(message.c_str(), std::regex("^VDisk is "), "VDisks are ");
-                            message = std::regex_replace(message.c_str(), std::regex("^VDisk "), "VDisks ");
-                            break;
-                        }
-                        case ETags::PDiskState: {
-                            message = std::regex_replace(message.c_str(), std::regex("^PDisk has "), "PDisks have ");
-                            message = std::regex_replace(message.c_str(), std::regex("^PDisk is "), "PDisks are ");
-                            message = std::regex_replace(message.c_str(), std::regex("^PDisk "), "PDisks ");
-                            break;
-                        }
-                        default:
-                            break;
-                    }
-                    it->IssueLog.set_message(message);
-                }
-            }
-        }
-
-        void FillRecords(TList<TSelfCheckResult::TIssueRecord>& records) {
-            for(auto it = recordsMap.begin(); it != recordsMap.end(); ++it) {
-                records.splice(records.end(), it->second);
-            }
-            RemoveUnlinkIssues(records);
-            RenameMergingIssues(records);
-        }
-
-        TList<TSelfCheckResult::TIssueRecord>& GetRecords(ETags tag) {
-            return recordsMap[tag];
-        }
-    };
-
-    bool FindRecordsForMerge(TList<TSelfCheckResult::TIssueRecord>& records, TList<TSelfCheckResult::TIssueRecord>& similar, TList<TSelfCheckResult::TIssueRecord>& merged) {
-        while (!records.empty() && similar.empty()) {
-            similar.splice(similar.end(), records, records.begin());
-            for (auto it = records.begin(); it != records.end(); ) {
-                bool isSimilar = it->IssueLog.status() == similar.begin()->IssueLog.status()
-                    && it->IssueLog.message() == similar.begin()->IssueLog.message()
-                    && it->IssueLog.level() == similar.begin()->IssueLog.level() ;
-                if (isSimilar && similar.begin()->Tag == ETags::VDiskState) {
-                    isSimilar = it->IssueLog.location().storage().node().id() == similar.begin()->IssueLog.location().storage().node().id();
-                }
-                if (isSimilar) {
-                    auto move = it++;
-                    similar.splice(similar.end(), records, move);
-                } else {
-                    ++it;
-                }
-            }
-
-            if (similar.size() == 1) {
-                merged.splice(merged.end(), similar);
-            }
-        }
-
-        return !similar.empty();
-    }
-
-    std::shared_ptr<TList<TSelfCheckResult::TIssueRecord>> FindChildrenRecords(TList<TSelfCheckResult::TIssueRecord>& records, TSelfCheckResult::TIssueRecord& parent) {
-        std::shared_ptr<TList<TSelfCheckResult::TIssueRecord>> children(new TList<TSelfCheckResult::TIssueRecord>);
-        std::unordered_set<TString> childrenIds;
-        for (auto reason: parent.IssueLog.reason()) {
-            childrenIds.insert(reason);
-        }
-
-        for (auto it = records.begin(); it != records.end(); ) {
-            if (childrenIds.contains(it->IssueLog.id())) {
-                auto move = it++;
-                children->splice(children->end(), records, move);
-            } else {
-                it++;
-            }
-        }
-
-        return children;
-    }
-
-    void MoveDataInFirstRecord(TMergeIssuesContext& context, TList<TSelfCheckResult::TIssueRecord>& similar) {
-        auto mainReasons = similar.begin()->IssueLog.mutable_reason();
-        std::unordered_set<TString> ids;
-        ids.insert(similar.begin()->IssueLog.id());
-        std::unordered_set<TString> mainReasonIds;
-        for (auto it = mainReasons->begin(); it != mainReasons->end(); it++) {
-            mainReasonIds.insert(*it);
-        }
-
-        for (auto it = std::next(similar.begin(), 1); it != similar.end(); ) {
-            if (ids.contains(it->IssueLog.id())) {
-                it++;
-                continue;
-            }
-            ids.insert(it->IssueLog.id());
-
-            switch (similar.begin()->Tag) {
-                case ETags::GroupState: {
-                    auto mainGroupIds = similar.begin()->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_id();
-                    auto donorGroupIds = it->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_id();
-                    mainGroupIds->Add(donorGroupIds->begin(), donorGroupIds->end());
-                    break;
-                }
-                case ETags::VDiskState: {
-                    auto mainVdiskIds = similar.begin()->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id();
-                    auto donorVdiskIds = it->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id();
-                    mainVdiskIds->Add(donorVdiskIds->begin(), donorVdiskIds->end());
-                    break;
-                }
-                case ETags::PDiskState: {
-                    auto mainPdisk = similar.begin()->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk();
-                    auto donorPdisk = it->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk();
-                    mainPdisk->Add(donorPdisk->begin(), donorPdisk->end());
-                    break;
-                }
-                default:
-                    break;
-            }
-
-            auto donorReasons = it->IssueLog.mutable_reason();
-            for (auto donorReasonIt = donorReasons->begin(); donorReasonIt != donorReasons->end(); donorReasonIt++) {
-                if (!mainReasonIds.contains(*donorReasonIt)) {
-                    mainReasons->Add(donorReasonIt->c_str());
-                    mainReasonIds.insert(*donorReasonIt);
-                }
-            }
-
-            context.removeIssuesIds.insert(it->IssueLog.id());
-            it = similar.erase(it);
-        }
-
-        similar.begin()->IssueLog.set_count(ids.size());
-        similar.begin()->IssueLog.set_listed(ids.size());
-    }
-
-    void MergeLevelRecords(TMergeIssuesContext& context, TList<TSelfCheckResult::TIssueRecord>& records) {
-        TList<TSelfCheckResult::TIssueRecord> handled;
-        while (!records.empty()) {
-            TList<TSelfCheckResult::TIssueRecord> similar;
-            if (FindRecordsForMerge(records, similar, handled)) {
-                MoveDataInFirstRecord(context, similar);
-                handled.splice(handled.end(), similar, similar.begin());
-            }
-        }
-        records.splice(records.end(), handled);
-    }
-
-    void MergeLevelRecords(TMergeIssuesContext& context, ETags levelTag) {
-        auto& records = context.GetRecords(levelTag);
-        MergeLevelRecords(context, records);
-    }
-
-    void MergeLevelRecords(TMergeIssuesContext& context, ETags levelTag, ETags upperTag) {
-        auto& levelRecords = context.GetRecords(levelTag);
-        auto& upperRecords = context.GetRecords(upperTag);
-
-        for (auto it = upperRecords.begin(); it != upperRecords.end(); it++) {
-            auto children = FindChildrenRecords(levelRecords, *it);
-            if (children->size() > 1) {
-                MergeLevelRecords(context, *children);
-            }
-            levelRecords.splice(levelRecords.end(), *children);
-        }
-    }
-
-    int GetIssueCount(TSelfCheckResult::TIssueRecord& record) {
-        return record.IssueLog.count() == 0 ? 1 : record.IssueLog.count();
-    }
-
-    void SetIssueCount(TSelfCheckResult::TIssueRecord& record, int value) {
-        if (record.IssueLog.listed() == 0) {
-            record.IssueLog.set_listed(1);
-        }
-        record.IssueLog.set_count(value);
-    }
-
-    int GetIssueListed(TSelfCheckResult::TIssueRecord& record) {
-        return record.IssueLog.listed() == 0 ? 1 : record.IssueLog.listed();
-    }
-
-    void SetIssueListed(TSelfCheckResult::TIssueRecord& record, int value) {
-        if (record.IssueLog.count() == 0) {
-            record.IssueLog.set_count(1);
-        }
-        record.IssueLog.set_listed(value);
-    }
-
-    void FillGroupStatus(TGroupId groupId, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckResult context) {
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_id()->Clear();
-        context.Location.mutable_storage()->mutable_pool()->mutable_group()->add_id(ToString(groupId));
-        storageGroupStatus.set_id(ToString(groupId));
-
-        auto itGroup = GroupState.find(groupId);
-        if (itGroup == GroupState.end()) {
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "Group has no vslots", ETags::GroupState);
-            storageGroupStatus.set_overall(context.GetOverallStatus());
-            return;
-        }
-
-        TGroupChecker checker(itGroup->second.ErasureSpecies, itGroup->second.LayoutCorrect);
-        const auto& slots = itGroup->second.VSlots;
-        for (const auto* slot : slots) {
-            const auto& slotInfo = slot->GetInfo();
-            auto slotId = GetVSlotId(slot->GetKey());
-            auto [statusIt, inserted] = VDiskStatuses.emplace(slotId, Ydb::Monitoring::StatusFlag::UNSPECIFIED);
-            if (inserted) {
-                Ydb::Monitoring::StorageVDiskStatus& vDiskStatus = *storageGroupStatus.add_vdisks();
-                FillVDiskStatus(slot, vDiskStatus, {&context, "VDISK"});
-                statusIt->second = vDiskStatus.overall();
-            }
-            checker.AddVDiskStatus(statusIt->second, slotInfo.GetFailRealm());
-        }
-
-        context.Location.mutable_storage()->clear_node(); // group doesn't have node
-        context.OverallStatus = MinStatus(context.OverallStatus, Ydb::Monitoring::StatusFlag::YELLOW);
-        checker.ReportStatus(context);
-
-        storageGroupStatus.set_overall(context.GetOverallStatus());
-    }
-
-    void MergeRecords(TList<TSelfCheckResult::TIssueRecord>& records) {
-        TMergeIssuesContext mergeContext(records);
-        if (Request->Request.merge_records()) {
-            MergeLevelRecords(mergeContext, ETags::GroupState);
-            MergeLevelRecords(mergeContext, ETags::VDiskState, ETags::GroupState);
-            MergeLevelRecords(mergeContext, ETags::PDiskState, ETags::VDiskState);
-        }
-        mergeContext.FillRecords(records);
-    }
-
-    void FillPoolStatus(const TStoragePoolState& pool, Ydb::Monitoring::StoragePoolStatus& storagePoolStatus, TSelfCheckResult context) {
-        context.Location.mutable_storage()->mutable_pool()->set_name(pool.Name);
-        storagePoolStatus.set_id(pool.Name);
-        bool haveInfo = HaveAllBSControllerInfo();
-        for (auto groupId : pool.Groups) {
-            if (haveInfo && !UnknownStaticGroups.contains(groupId)) {
-                FillGroupStatus(groupId, *storagePoolStatus.add_groups(), {&context, "STORAGE_GROUP"});
-            } else if (IsStaticGroup(groupId)) {
-                auto itGroup = MergedBSGroupState.find(groupId);
-                if (itGroup != MergedBSGroupState.end()) {
-                    FillGroupStatusWithWhiteboard(groupId, *itGroup->second, *storagePoolStatus.add_groups(), {&context, "STORAGE_GROUP"});
-                }
-            }
-        }
-
-        MergeRecords(context.IssueRecords);
-
-        switch (context.GetOverallStatus()) {
-            case Ydb::Monitoring::StatusFlag::BLUE:
-            case Ydb::Monitoring::StatusFlag::YELLOW:
-                context.ReportStatus(context.GetOverallStatus(), "Pool degraded", ETags::PoolState, {ETags::GroupState});
-                break;
-            case Ydb::Monitoring::StatusFlag::ORANGE:
-                context.ReportStatus(context.GetOverallStatus(), "Pool has no redundancy", ETags::PoolState, {ETags::GroupState});
-                break;
-            case Ydb::Monitoring::StatusFlag::RED:
-                context.ReportStatus(context.GetOverallStatus(), "Pool failed", ETags::PoolState, {ETags::GroupState});
-                break;
-            default:
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                break;
-        }
-        storagePoolStatus.set_overall(context.GetOverallStatus());
-    }
-
-    void FillStorage(TDatabaseState& databaseState, Ydb::Monitoring::StorageStatus& storageStatus, TSelfCheckResult context) {
-        if (HaveAllBSControllerInfo() && databaseState.StoragePools.empty()) {
-            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "There are no storage pools", ETags::StorageState);
-        } else {
-            if (HaveAllBSControllerInfo()) {
-                for (const ui64 poolId : databaseState.StoragePools) {
-                    auto itStoragePoolState = StoragePoolState.find(poolId);
-                    if (itStoragePoolState != StoragePoolState.end() && itStoragePoolState->second.Groups) {
-                        FillPoolStatus(itStoragePoolState->second, *storageStatus.add_pools(), {&context, "STORAGE_POOL"});
-                        StoragePoolSeen.emplace(poolId);
-                    }
-                }
-            } else {
-                for (const TString& poolName : databaseState.StoragePoolNames) {
-                    auto itStoragePoolState = StoragePoolStateByName.find(poolName);
-                    if (itStoragePoolState != StoragePoolStateByName.end()) {
-                        FillPoolStatus(itStoragePoolState->second, *storageStatus.add_pools(), {&context, "STORAGE_POOL"});
-                        StoragePoolSeenByName.emplace(poolName);
-                    }
-                }
-            }
-            switch (context.GetOverallStatus()) {
-                case Ydb::Monitoring::StatusFlag::BLUE:
-                case Ydb::Monitoring::StatusFlag::YELLOW:
-                    context.ReportStatus(context.GetOverallStatus(), "Storage degraded", ETags::StorageState, {ETags::PoolState});
-                    break;
-                case Ydb::Monitoring::StatusFlag::ORANGE:
-                    context.ReportStatus(context.GetOverallStatus(), "Storage has no redundancy", ETags::StorageState, {ETags::PoolState});
-                    break;
-                case Ydb::Monitoring::StatusFlag::RED:
-                    context.ReportStatus(context.GetOverallStatus(), "Storage failed", ETags::StorageState, {ETags::PoolState});
-                    break;
-                default:
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
-                    break;
-            }
-            if (!HaveAllBSControllerInfo()) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet BSC didn't provide information", ETags::StorageState);
-            }
-        }
-        if (databaseState.StorageQuota > 0) {
-            auto usage = (float)databaseState.StorageUsage / databaseState.StorageQuota;
-            if (usage > 0.9) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Storage usage over 90%", ETags::StorageState);
-            } else if (usage > 0.85) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Storage usage over 85%", ETags::StorageState);
-            } else if (usage > 0.75) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Storage usage over 75%", ETags::StorageState);
-            }
-        }
-        storageStatus.set_overall(context.GetOverallStatus());
-    }
-
-    struct TOverallStateContext {
-        Ydb::Monitoring::SelfCheckResult* Result;
-        Ydb::Monitoring::StatusFlag::Status Status = Ydb::Monitoring::StatusFlag::GREY;
-        bool HasDegraded = false;
-        std::unordered_set<std::pair<TString, TString>> IssueIds;
-
-        TOverallStateContext(Ydb::Monitoring::SelfCheckResult* result) {
-            Result = result;
-        }
-
-        void FillSelfCheckResult() {
-            switch (Status) {
-            case Ydb::Monitoring::StatusFlag::GREEN:
-                Result->set_self_check_result(Ydb::Monitoring::SelfCheck::GOOD);
-                break;
-            case Ydb::Monitoring::StatusFlag::YELLOW:
-                if (HasDegraded) {
-                    Result->set_self_check_result(Ydb::Monitoring::SelfCheck::DEGRADED);
-                } else {
-                    Result->set_self_check_result(Ydb::Monitoring::SelfCheck::GOOD);
-                }
-                break;
-            case Ydb::Monitoring::StatusFlag::BLUE:
-                Result->set_self_check_result(Ydb::Monitoring::SelfCheck::DEGRADED);
-                break;
-            case Ydb::Monitoring::StatusFlag::ORANGE:
-                Result->set_self_check_result(Ydb::Monitoring::SelfCheck::MAINTENANCE_REQUIRED);
-                break;
-            case Ydb::Monitoring::StatusFlag::RED:
-                Result->set_self_check_result(Ydb::Monitoring::SelfCheck::EMERGENCY);
-                break;
-            default:
-                break;
-            }
-        }
-
-        void UpdateMaxStatus(Ydb::Monitoring::StatusFlag::Status status) {
-            Status = MaxStatus(Status, status);
-        }
-
-        void AddIssues(TList<TSelfCheckResult::TIssueRecord>& issueRecords) {
-            for (auto& issueRecord : issueRecords) {
-                std::pair<TString, TString> key{issueRecord.IssueLog.location().database().name(), issueRecord.IssueLog.id()};
-                if (IssueIds.emplace(key).second) {
-                    Result->mutable_issue_log()->Add()->CopyFrom(issueRecord.IssueLog);
-                }
-            }
-        }
-    };
-
-    void FillDatabaseResult(TOverallStateContext& context, const TString& path, TDatabaseState& state) {
-        Ydb::Monitoring::DatabaseStatus& databaseStatus(*context.Result->add_database_status());
-        TSelfCheckResult dbContext("DATABASE");
-        if (!IsSpecificDatabaseFilter()) {
-            dbContext.Location.mutable_database()->set_name(path);
-        }
-        databaseStatus.set_name(path);
-        FillCompute(state, *databaseStatus.mutable_compute(), {&dbContext, "COMPUTE"});
-        FillStorage(state, *databaseStatus.mutable_storage(), {&dbContext, "STORAGE"});
-        if (databaseStatus.compute().overall() != Ydb::Monitoring::StatusFlag::GREEN
-                && databaseStatus.storage().overall() != Ydb::Monitoring::StatusFlag::GREEN) {
-            dbContext.ReportStatus(MaxStatus(databaseStatus.compute().overall(), databaseStatus.storage().overall()),
-                "Database has multiple issues", ETags::DBState, { ETags::ComputeState, ETags::StorageState});
-        } else if (databaseStatus.compute().overall() != Ydb::Monitoring::StatusFlag::GREEN) {
-            dbContext.ReportStatus(databaseStatus.compute().overall(), "Database has compute issues", ETags::DBState, {ETags::ComputeState});
-        } else if (databaseStatus.storage().overall() != Ydb::Monitoring::StatusFlag::GREEN) {
-            dbContext.ReportStatus(databaseStatus.storage().overall(), "Database has storage issues", ETags::DBState, {ETags::StorageState});
-        }
-        databaseStatus.set_overall(dbContext.GetOverallStatus());
-        context.UpdateMaxStatus(dbContext.GetOverallStatus());
-        context.AddIssues(dbContext.IssueRecords);
-        if (!context.HasDegraded && context.Status != Ydb::Monitoring::StatusFlag::GREEN && dbContext.HasTags({ETags::StorageState})) {
-            context.HasDegraded = true;
-        }
-    }
-
-    void FillResult(TOverallStateContext context) {
-        if (IsSpecificDatabaseFilter()) {
-            FillDatabaseResult(context, FilterDatabase, DatabaseState[FilterDatabase]);
-        } else {
-            for (auto& [path, state] : DatabaseState) {
-                FillDatabaseResult(context, path, state);
-            }
-        }
-        if (DatabaseState.empty()) {
-            Ydb::Monitoring::DatabaseStatus& databaseStatus(*context.Result->add_database_status());
-            TSelfCheckResult tabletContext;
-            tabletContext.Location.mutable_database()->set_name(DomainPath);
-            databaseStatus.set_name(DomainPath);
-            {
-                TDatabaseState databaseState;
-                FillSystemTablets(databaseState, {&tabletContext, "SYSTEM_TABLET"});
-                context.UpdateMaxStatus(tabletContext.GetOverallStatus());
-            }
-        }
-        if (!FilterDatabase) {
-            TDatabaseState unknownDatabase;
-            bool fillUnknownDatabase = false;
-            if (HaveAllBSControllerInfo()) {
-                for (auto& [id, pool] : StoragePoolState) {
-                    if (StoragePoolSeen.count(id) == 0) {
-                        unknownDatabase.StoragePools.insert(id);
-                    }
-                }
-                fillUnknownDatabase = !unknownDatabase.StoragePools.empty();
-            } else {
-                for (auto& [name, pool] : StoragePoolStateByName) {
-                    if (StoragePoolSeenByName.count(name) == 0) {
-                        unknownDatabase.StoragePoolNames.insert(name);
-                    }
-                }
-                fillUnknownDatabase = !unknownDatabase.StoragePoolNames.empty();
-            }
-            if (fillUnknownDatabase) {
-                Ydb::Monitoring::DatabaseStatus& databaseStatus(*context.Result->add_database_status());
-                TSelfCheckResult storageContext;
-                FillStorage(unknownDatabase, *databaseStatus.mutable_storage(), {&storageContext, "STORAGE"});
-                databaseStatus.set_overall(storageContext.GetOverallStatus());
-                context.UpdateMaxStatus(storageContext.GetOverallStatus());
-                context.AddIssues(storageContext.IssueRecords);
-            }
-        }
-        context.FillSelfCheckResult();
-    }
-
     bool TruncateIssuesWithStatusWhileBeyondLimit(Ydb::Monitoring::SelfCheckResult& result, ui64 byteLimit, Ydb::Monitoring::StatusFlag::Status status) {
         auto byteResult = result.ByteSizeLong();
         if (byteResult <= byteLimit) {
@@ -2950,10 +1414,47 @@ public:
         }
     }
 
+    THolder<TBuilderResult<TBuilderDatabase, TBuilderSystemTablets, TBuilderStorage>> MakeBuilder() {
+        auto builderStoragePDiskWB = MakeHolder<TBuilderStoragePDiskWithWhiteboard>();
+        auto builderStorageVDiskWB = MakeHolder<TBuilderStorageVDiskWithWhiteboard>(builderStoragePDiskWB);
+        auto builderStorageGroupWB = MakeHolder<TBuilderStorageGroupWithWhiteboard>(builderStorageVDiskWB);
+
+        auto builderStoragePDisk = MakeHolder<TBuilderStoragePDisk>(builderStoragePDiskWB);
+        auto builderStorageVDisk = MakeHolder<TBuilderStorageVDisk>(builderStorageVDiskWB);
+        auto builderStorageGroup = MakeHolder<TBuilderStorageGroup>(builderStorageGroupWB);
+
+        auto builderStoragePool = MakeHolder<TBuilderStoragePool>(builderStorageGroup);
+        auto builderStorage = MakeHolder<TBuilderStorage>(builderStoragePool);
+
+        auto builderComputeNode = MakeHolder<TBuilderComputeNode>();
+        auto builderComputeDatabase = MakeHolder<TBuilderComputeDatabase>(builderComputeNode);
+        auto builderCompute = MakeHolder<TBuilderCompute>(builderComputeDatabase, FilterDomainKey, MergedNodeSystemState);
+
+        auto builderDatabase = MakeHolder<TBuilderDatabase>(builderCompute, builderStorage, IsSpecificDatabaseFilter());
+
+        // root builder
+        auto builder = MakeHolder<TBuilderResult<TBuilderDatabase, TBuilderSystemTablets, TBuilderStorage>>(
+            DomainPath, FilterDatabase, DatabaseState,
+            StoragePoolState, StoragePoolSeen,
+            StoragePoolStateByName, StoragePoolSeenByName,
+            builderDatabase,
+            builderCompute,
+            builderStorage,
+            builderStoragePDisk,
+            builderStorageVDisk,
+            builderStorageGroup,
+            builderStoragePDiskWB,
+            builderStorageVDiskWB,
+            builderStorageGroupWB
+        );
+
+        return builder;
+    }
+
     void ReplyAndPassAway() {
         Span.Event("ReplyAndPassAway");
         THolder<TEvSelfCheckResult> response = MakeHolder<TEvSelfCheckResult>();
-        Ydb::Monitoring::SelfCheckResult& result = response->Result;
+        Ydb::Monitoring::SelfCheckResult& result = builder;
 
         AggregateHiveInfo();
         AggregateHiveNodeStats();
@@ -2964,7 +1465,8 @@ public:
             TabletRequests.TabletStates[tabletId].IsUnresponsive = true;
         }
 
-        FillResult({&result});
+        auto builder = MakeBuilder();
+        builder.Result.Build({&result});
         RemoveUnrequestedEntries(result, Request->Request);
 
         FillNodeInfo(SelfId().NodeId(), result.mutable_location());
@@ -3003,6 +1505,51 @@ public:
         }
         PassAway();
     }
+
+    class TSelfCheckBuilder {
+public:
+    THolder<IBuilderResult> Result;
+    THolder<IBuilderDatabase> Database;
+    THolder<IBuilderCompute> Compute;
+    THolder<IBuilderComputeDatabase> ComputeDatabase;
+    THolder<IBuilderComputeNode> ComputeNode;
+    THolder<IBuilderStorage> Storage;
+    THolder<IBuilderStoragePool> StoragePool;
+    THolder<IBuilderStorageGroup> StorageGroup;
+    THolder<IBuilderStorageVDisk> StorageVDisk;
+    THolder<IBuilderStoragePDisk> StoragePDisk;
+    THolder<IBuilderStorageGroupWithWhiteboard> StorageGroupWithWhiteboard;
+    THolder<IBuilderStorageVDiskWithWhiteboard> StorageVDiskWithWhiteboard;
+    THolder<IBuilderStoragePDiskWithWhiteboard> StoragePDiskWithWhiteboard;
+
+    void Init(THolder<IBuilderResult> result,
+                      THolder<IBuilderDatabase> database,
+                      THolder<IBuilderCompute> compute,
+                      THolder<IBuilderComputeDatabase> computeDatabase,
+                      THolder<IBuilderComputeNode> computeNode,
+                      THolder<IBuilderStorage> storage,
+                      THolder<IBuilderStoragePool> storagePool,
+                      THolder<IBuilderStorageGroup> storageGroup,
+                      THolder<IBuilderStorageGroupWithWhiteboard> storageGroupWithWhiteboard,
+                      THolder<IBuilderStoragePDisk> storagePDisk,
+                      THolder<IBuilderStoragePDiskWithWhiteboard> storagePDiskWithWhiteboard,
+                      THolder<IBuilderStorageVDisk> storageVDisk,
+                      THolder<IBuilderStorageVDiskWithWhiteboard> storageVDiskWithWhiteboard) {
+        Result = std::move(result);
+        Database = std::move(database);
+        Compute = std::move(compute);
+        ComputeDatabase = std::move(computeDatabase);
+        ComputeNode = std::move(computeNode);
+        Storage = std::move(storage);
+        StoragePool = std::move(storagePool);
+        StorageGroup = std::move(storageGroup);
+        StorageGroupWithWhiteboard = std::move(storageGroupWithWhiteboard);
+        StoragePDisk = std::move(storagePDisk);
+        StoragePDiskWithWhiteboard = std::move(storagePDiskWithWhiteboard);
+        StorageVDisk = std::move(storageVDisk);
+        StorageVDiskWithWhiteboard = std::move(storageVDiskWithWhiteboard);
+    }
+};
 };
 
 } // NKikimr::NHealthCheck
